pyproject.toml
[project]
name = "dhurba-furniture-agent"
version = "0.1.0"
description = "Langgraph agent for Dhurba Furniture"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.1.4",
    "langgraph>=0.4.5",
    "python-dotenv>=1.1.0",
    "requests>=2.31.0",
    "sqlalchemy>=2.0.0",
    "pandas>=2.0.0",
    "pinecone>=6.0.1",
    "langchain-pinecone>=0.2.6",
    "sentence-transformers>=4.1.0",
    "langchain-huggingface>=0.2.0",
    "langgraph-api>=0.2.42",
]
[tool.setuptools]
packages = ["agent"]
[dependency-groups]
dev=[
    "ipykernel>=6.29.5",
    "langgraph-cli>=0.2.10",
]
